DELIA-56463: Observed RDK-03088 while connecting XRE
Reason for change: Xre is trying to connect via ipv6 forecully though we didn't
receive ipv6 global address which causes some delay in xre connection. So, created a
file immediately after receiving global address and add a check to ensure
whether global address is received or not before sending event to receiver.
Test Procedure: There should not be any delay during XRE connection
Risks: High

(cherry picked from commit 7c040c3c163c7a41bc13c186360814745f54afeb)

+++ ipmodechange.sh
@@ -49,30 +49,38 @@ if [ "$opern" = "add" ]; then
 	   touch /tmp/ipv4_route
 	   if [ ! -f /tmp/ipv6_mode ] && [ ! -f /tmp/ipv4_mode ];
 	   then
+               if [ -f /tmp/ipv4_global ]; then
 		   echo "Publishing ipmode change to ipv4" >> $LOGFILE
 		   /usr/bin/IARM_event_sender IpmodeEvent 1 ipv4
 		   touch /tmp/ipv4_mode
+               fi
 	   fi
    elif [ "x$lmode" = "xipv6" ]; then
 	   touch /tmp/ipv6_route
-	   echo "Publishing ipmode change to ipv6" >> $LOGFILE
-	   /usr/bin/IARM_event_sender IpmodeEvent 1 ipv6
-	   touch /tmp/ipv6_mode
+           if [ -f /tmp/ipv6_global ]; then
+               echo "Publishing ipmode change to ipv6" >> $LOGFILE
+               /usr/bin/IARM_event_sender IpmodeEvent 1 ipv6
+               touch /tmp/ipv6_mode
+           fi
    fi
 elif [ "$opern" = "delete" ]; then
     #Remove flag and IP for delete operation
     echo "Deleting Route Flag" >> $LOGFILE
     if [ "x$lmode" = "xipv6" ];
     then
-	   rm -f  /tmp/ipv6_mode
 	   rm -f  /tmp/ipv6_route
-	   if [ -f /tmp/ipv4_route ]; then
+           if [ ! -f /tmp/ipv6_global ]; then
+               rm -f  /tmp/ipv6_mode
+               if [ -f /tmp/ipv4_route ]; then
 		   echo "Publishing ipmode change to ipv4" >> $LOGFILE
 		   /usr/bin/IARM_event_sender IpmodeEvent 1 ipv4
 		   touch /tmp/ipv4_mode
+               fi
 	   fi
    elif [ "x$lmode" = "xipv4" ]; then
-	   rm -f /tmp/ipv4_mode
+           if [ ! -f /tmp/ipv4_global ]; then
+               rm -f /tmp/ipv4_mode
+           fi
 	   rm -f  /tmp/ipv4_route
    fi
 else
